Extract test control server to separate module
Move startTestControlServer from index.ts to test-control-server.ts
to keep the main entry point minimal.
