+++ src/main/index.ts
@@ -3,44 +3,7 @@ import { AppWindowManager } from './browser/app-window-manager';
 import { DataStoreManager } from './database/data-store-manager';
 import { DownloadManager } from './browser/download-manager';
 import { SettingsEnforcer } from './settings/settings-enforcer';
-
-// Lightweight test control server for perf tests.
-// Activated when REMOTE_DEBUGGING_PORT env var is set.
-function startTestControlServer(port: number): void {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const http = require('http') as typeof import('http');
-  const server = http.createServer((req: import('http').IncomingMessage, res: import('http').ServerResponse) => {
-    res.setHeader('Content-Type', 'application/json');
-
-    if (req.method === 'GET' && req.url === '/status') {
-      const win = AppWindowManager.getActiveWindow();
-      res.end(JSON.stringify({ ready: true, hasWindow: !!win }));
-      return;
-    }
-
-    if (req.method === 'POST' && req.url?.startsWith('/create-tab')) {
-      const parsed = new URL(req.url, `http://localhost:${port}`);
-      const tabUrl = parsed.searchParams.get('url') || 'about:blank';
-      const win = AppWindowManager.getActiveWindow();
-      if (win) {
-        win.createTab(tabUrl, true).then(() => {
-          res.end(JSON.stringify({ ok: true }));
-        }).catch((err: Error) => {
-          res.statusCode = 500;
-          res.end(JSON.stringify({ error: err.message }));
-        });
-      } else {
-        res.statusCode = 500;
-        res.end(JSON.stringify({ error: 'no active window' }));
-      }
-      return;
-    }
-
-    res.statusCode = 404;
-    res.end(JSON.stringify({ error: 'not found' }));
-  });
-  server.listen(port, '127.0.0.1');
-}
+import { startTestControlServer } from './test-control-server';
 
 // Disable Chromium features that trigger macOS "Local Network" permission dialog.
 // These features use mDNS/Bonjour for device discovery, which is unnecessary for
